fix: address Humanizer migration findings — i64 support, MCP improvements
Compiler/Language fixes:
- Fix integer overflow data corruption: widen IntLiteralNode and
  BoundIntLiteral from int to long, preventing silent truncation of
  C# long literals during C#→Calor conversion
- Lexer: add long.TryParse fallback for values outside int range
- Lexer: support numeric separators (1_000_000) by consuming underscores
- Lexer: fix hex literal truncation — store as long when > int.MaxValue
- CSharpEmitter: add L suffix for values outside int range
- RoslynSyntaxVisitor: remove (int)longVal cast that caused data corruption

MCP improvements:
- McpServer: set Console.InputEncoding/OutputEncoding to UTF-8 in
  CreateStdio() to fix § character corruption on Windows
- BatchConvertTool: return structured validation errors with line,
  column, category, and suggestion fields instead of flat strings
- CompileTool: add errorCategories summary to batch compilation output
  showing error counts grouped by error code
- DiagnoseTool: add 5 new error-suggestion patterns for common
  converter output issues (§NEW{object}, integer overflow, generic
  type params, IF ID mismatch)

+++ src/Calor.Compiler/Analysis/BugPatterns/Patterns/OverflowChecker.cs
@@ -253,7 +253,7 @@ public sealed class OverflowChecker : IBugPatternChecker
                Math.Abs(rightValue.Value) < SmallThreshold;
     }
 
-    private static int? GetConstantValue(BoundExpression expr)
+    private static long? GetConstantValue(BoundExpression expr)
     {
         return expr switch
         {
@@ -279,7 +279,7 @@ public sealed class OverflowChecker : IBugPatternChecker
                         BinaryOperator.Add => leftValue.Value + rightValue.Value,
                         BinaryOperator.Subtract => leftValue.Value - rightValue.Value,
                         BinaryOperator.Multiply => leftValue.Value * rightValue.Value,
-                        BinaryOperator.LeftShift => leftValue.Value << rightValue.Value,
+                        BinaryOperator.LeftShift => leftValue.Value << (int)rightValue.Value,
                         _ => 0
                     };
                 }
